backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
supabase>=2.27.0
sqlalchemy==2.0.23

# Machine Learning
torch>=2.2.0
torchvision>=0.17.0
pillow>=10.1.0

# MLOps
mlflow==2.9.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx>=0.28.0
hypothesis>=6.92.0
websockets>=15.0
hypothesis>=6.92.0

# Data Science
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.66.0
